code/requirements.txt
sanic
requests
Pillow==9.4.0
traceback2
python-math 
torch
torchvision==0.12.0
numpy==1.19.2
albumentations==0.4.3
datasets==2.8.0
diffusers
opencv-python==4.1.2.30
pudb==2019.2
invisible-watermark
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.4.2
omegaconf==2.1.1
test-tube>=0.7.5
streamlit>=0.73.1
einops==0.3.0
torch-fidelity==0.3.0
transformers
torchmetrics==0.6.0
kornia==0.6
-e git+https://github.com/CompVis/taming-transformers.git@master#egg=taming-transformers
-e git+https://github.com/openai/CLIP.git@main#egg=clip
huggingface-hub
openai
seaborn
git+https://github.com/crowsonkb/k-diffusion.git
git+https://github.com/huggingface/diffusers